+++ Helper/tinyxml2.h
@@ -289,7 +289,7 @@ private:
         if ( cap > _allocated ) {
             int newAllocated = cap * 2;
             T* newMem = new T[newAllocated];
-            memcpy( newMem, _mem, sizeof(T)*_size );	// warning: not using constructors, only works for PODs
+            memcpy( newMem, _mem, (long unsigned int)(sizeof(T)*_size) );	// warning: not using constructors, only works for PODs
             if ( _mem != _pool ) {
                 delete [] _mem;
             }
